Update native code id to be 8 bytes from 1byte

+++ radix-engine-interface/src/blueprints/package/invocations.rs
@@ -57,7 +57,7 @@ pub const PACKAGE_PUBLISH_NATIVE_IDENT: &str = "publish_native";
 #[derive(Debug, Clone, Eq, PartialEq, ScryptoSbor)]
 pub struct PackagePublishNativeInput {
     pub package_address: Option<GlobalAddressReservation>,
-    pub native_package_code_id: u8,
+    pub native_package_code_id: u64,
     pub setup: PackageDefinition,
     pub metadata: BTreeMap<String, MetadataValue>,
 }
@@ -65,7 +65,7 @@ pub struct PackagePublishNativeInput {
 #[derive(Debug, Clone, Eq, PartialEq, ManifestSbor)]
 pub struct PackagePublishNativeManifestInput {
     pub package_address: Option<ManifestAddressReservation>,
-    pub native_package_code_id: u8,
+    pub native_package_code_id: u64,
     pub setup: PackageDefinition,
     pub metadata: BTreeMap<String, MetadataValue>,
 }
